[udev] Suspend proof event reading and processing. Fixes JB#34440
Depending on kernel side usb drivers and charging logic, it is possible
that device suspends immediately after reporting usb connect / disconnect
while display is off and autosuspend enabled. That can delay the actions
that should be made on usb connection state changes.

Acquire a wakelock for the duration of udev event input processing.

Hold a wakelock for a while when usb connection state changes occur.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>