[config] Reverse the meaning of config_value_changed() return value
Having a predicate function with name xxx_changed() that returns a
non-zero value when nothing has changed is confusing.

Switch the return value meaning so that non-zero means there is a
difference and zero that there is not.

Also treat old value of NULL and new value of NULL as no change.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>