[udev] Fix udev iowatch issues
The io watch source id is not cleared when it is going to be implicitly
removed - which can lead to unsuccessful removal attempt when/if
hwal_cleanup() function is called later on.

Clear watch_id when return value from the callback function is going
to cause it to be implicitly removed.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>