[appsync] Start/stop only applications relevant to mode. JB#35683
The boolean state data for tracking active/inactive applications makes
it difficult to separate applications that are not relevant to the mode
and as a result an attempt to stop also those irrelevant apps that were
not started is made when exiting the mode.

Switch from boolean to inactive|active|dontcare state tracking.

When starting appsync, mark apps relevant to the mode as inactive and
the rest as dontcare.

When stopping appsync, stop only those applications that have been
activated by usb-moded.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>