[usb-moded] Do not force appsync stop during bootup. MER#1694
Starting/stopping appsync related systemd units is relatively expensive
task to perform and can cause timing problems during bootup.

As we know the appsync processes have not been started when device is
booting up, we can skip the explicit stopping when usb-moded is started
as part of bootup sequence.

Also, when usb-moded is exiting, it should be enough to stop only those
units that have been started by usb-moded.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>