[logging] Flush output streams relevant for logging
Mixed use of syslog, stdout, stderr and differences in buffering of
the streams causes the initial "starting up" message to end up as the
final output in usb-moded journal when debug level logging is used
(i.e. when freopen steps are skipped and stdout is not implicitly
flushed).

Use stderr and flush it before potentially switching to syslog logging.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>