[modules] Clear context pointer after dropping reference
Not explicitly initializing the pointer to null value makes the validity
of the initial state subject to compiler/libc level implementation details.
And not clearing the pointer after dropping the reference can lead to
hard to debug issues if the potentially state pointer is used later on.

Explicitly initialize the context to null value and set it back to null
after dropping the reference.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>