[modesetting] Track expected sysfs control file content. JB#38433
When porting to new devices it is easy to miss usb related control actions
taken for example by android init scripts and the resulting conflicts with
usb-moded logic might be difficult to spot and/or debug.

Keep track of sysfs values written by usb-moded and add functionality for
checking if something else has overruled the settings made by usb-moded.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>