[usb] Add a whitelist property specifying allowed USB modes. Contributes to JB#33745
If the whitelist property exists in the conf file only modes that
appear in the list can be used or set as the default.  If a current
mode is removed from the whitelist the device will be put into charging
mode and if the default is removed that will be changed to ask.