[modesetting] Write dummy value to clear ffs function list. Fixes JB#39402
It seems that kernels in many if not all devices ignore writing of empty
string to ffs function control file - the function list is left as it were
and this can cause subtle issues on the next cable connect. On the other
hand writing a non-existing "function" will clear the function list, but
causes write error to be reported.

Treat attempts to clear function list as a special case so that non-empty
string gets written and the resulting write error is ignored under normal
logging verbosity.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>