[udev] Suppress repetitive property warnings. Fixes JB#40394
Usb-moded expects to see POWER_SUPPLY_PRESENT and POWER_SUPPLY_TYPE
properties defined for usb power-supply devices. When this is not
the case, it emits warnings and applies various fallbacks. Since the
properties get re-evaluated periodically, this also leads to the
same warnings being repeated.

Refactor the code so that in default verbosity the property warnings
are emitted only during what looks like cable connect / disconnect.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>