[udev] Do not exit if cable state can't be determined
If usb-moded exits when it can't deduce cable connect state, it will
be restarted by systemd. As the usb-moded restarts are not going to
change how the udev status looks like, this repeats until systemd
moves the usb-moded service to failed state.

Do not exit and assume no usb-cable is connected. In addition to not
triggering restart spree, this has the additional benefit that the
usb-moded dbus interface stays available and signaling not-connected
state.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>