[android] Do not require android configuration block. Fixes JB#41303
Usb-moded does android usb initialization only if there is a non-empty
configuration block in the ini-files. This does not make sense anymore
as the values defined in the config file are used as fallback solution
if required data can't be obtained by other means.

If android_usb specific sysfs controls are present on the device, try
to initialize them regardless of whether ubs-moded configuration files
contain relevant information or not.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>