[worker] Use copy of target mode during transition. JB#43781
The common_map_mode_to_hardware() function can return the same
string that it was given as an argument. In the context of
worker_execute() function this can lead to using string pointer
that can get invalidated after releasing mutex.

Allocate copy of target mode before releasing mutex and use it
in the subsequent operations.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>