[worker] Stop mtp daemon on mode switch. JB#43781
If entry to mtp mode gets aborted due to cable disconnect, it is possible
that mtp daemon has already been started but it does not yet show up when
probing. This can lead to leaving stray/defunct mtp daemon instances
behind - which then can cause attempt to activate mtp mode after reconnect
to fail.

When leaving mode, stop mtp daemon if it seems to be running or we have
made an attempt to start it.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>