[logging] Add enablers for exposing call stack
When debugging complex issues it can be helpful know also which thread is
emitting the diagnostic message and the code path that led to the situation.

This is meant to be used as a development time debugging feature and thus
requires enabling during build time - otherwise all related code is excluded
from usb-moded binary.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>