[config] Ignore legacy mode=ask setting. JB#45869
For the purpose of preserving runtime modifications values in legacy
configuration files need to be given priority over new style static
configuration files. As a side effect this makes it impossible to
override default mode if it was done via defining mode=ask in the
legacy config file.

Ignore mode=ask if defined in a legacy config file.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>