[usb-moded] Make diagnostic mode selection permanent. JB#38342
Different config files must be used when diagnostic mode is
selected. In practice diagnostic mode selection is locked early
on during usb-moded startup and never changes. But as long as it
is possible, it would need to be taken into account in situations
where configuration files need to be re-evaluated.

Make it impossible to change diagnostic mode selection once it
has been made.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>