[appsync] Reload appsync configuration on SIGHUP. Fixes JB#38342
Developer mode does not work immediately after installing related
packages. Developer mode does become available for selecting, but
it does not work as expected because dhcpd is not started on
activation.

This is because usb-moded fails to reload appsync configuration
upon receiving SIGHUP and previous configuration state stays
effective until usb-moded is restarted (or device rebooted).

As appsync config contains state data that is used in worker
thread, existing configuration can't be just discarded and
replaced when processing SIGHUP from control thread.

Refactor appsync logic so that it keeps track of both active and
updated configuration. Load updated configuration on SIGHUP. Then
switch active configuration on the next mode transition - so that
the previous state is retained while it might be needed.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>