[udev] Optionally track extcon and android_usb devices too. JB#59336
In some devices charger power_supply device is not providing all
information usb-moded needs or is not sending change notifications,
which then can cause usb-moded not to react when cable is attached
or removed, prompt for usb mode when charger is attached, etc.

If device has functioning usb reporting via extcon subsystem, it can
be used to augment / override charger information by adding usb-moded
 configuration file with content like:

  [udev]
  extcon_tracking = 1

It is expected that there can be several extcon devices, and usb-moded
tracks all of them. Usually this should not be a problem, but if
conflicting usb state info is reported it might be necessary to
explicitly state which device to use:

  [udev]
  extcon_tracking = 1
  extcon_path = /sys/class/extcon/devicename

It is assumed that extcon devices report pc type usb connections. If
this is not the case and both charger and pc connections are reported,
information from android_usb subsystem devices can be used for
differentiating between charger and pc connections:

  [udev]
  extcon_tracking = 1
  android_tracking = 1

In some extreme cases it might be beneficial to make usb-moded ignore
charger device altogether. That can be accomplished e.g. via:

  [udev]
  charger_tracking = 0
  extcon_tracking = 1
  android_tracking = 1

By default usb-moded should behave just like before. But scoring
mechanism used for charger device selection heuristics was tweaked
a bit, so in some cases it might be necessary to explicitly state
which device should be tracked.

  [udev]
  path = /sys/class/power_supply/devicename

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>